Review cleanup: drop unused NetworkConditions monitor
